+++ src/compile/CompilerConfig.ts
@@ -1,6 +1,8 @@
-import { SourceResolver, SourcesMap, SourcesArray } from '@ton-community/func-js';
 import { Cell } from '@ton/core';
-import { Options } from '@tact-lang/compiler';
+
+import { TolkCompilerConfig } from './tolk/config';
+import { FuncCompilerConfig } from './func/config';
+import { TactCompilerConfig } from './tact/config';
 
 export type HookParams = {
     userData?: any;
@@ -35,34 +37,6 @@ export type CommonCompilerConfig = {
      * };
      */
     postCompileHook?: (code: Cell, params: HookParams) => Promise<void>;
-};
-
-export type TactCompilerConfig = {
-    lang: 'tact';
-    target: string;
-    options?: Options;
-};
-
-export type FuncCompilerConfig = {
-    lang?: 'func';
-    optLevel?: number;
-} & (
-    | {
-          targets: string[];
-          sources?: SourceResolver | SourcesMap;
-      }
-    | {
-          targets?: string[];
-          sources: SourcesArray;
-      }
-);
-
-export type TolkCompilerConfig = {
-    lang: 'tolk';
-    entrypoint: string;
-    optimizationLevel?: number;
-    withStackComments?: boolean;
-    experimentalOptions?: string;
 };
 
 export type CompilerConfig = (TactCompilerConfig | FuncCompilerConfig | TolkCompilerConfig) & CommonCompilerConfig;
