+++ src/compile/SourceSnapshot.ts
@@ -1,0 +1,4 @@
+export type SourceSnapshot = {
+    filename: string;
+    content: string;
+};
