+++ src/compile/func/config.ts
@@ -1,0 +1,15 @@
+import { SourceResolver, SourcesArray, SourcesMap } from '@ton-community/func-js';
+
+export type FuncCompilerConfig = {
+    lang?: 'func';
+    optLevel?: number;
+} & (
+    | {
+          targets: string[];
+          sources?: SourceResolver | SourcesMap;
+      }
+    | {
+          targets?: string[];
+          sources: SourcesArray;
+      }
+);
