+++ src/compile/tact/config.ts
@@ -1,0 +1,7 @@
+import {Options} from '@tact-lang/compiler';
+
+export type TactCompilerConfig = {
+    lang: 'tact';
+    target: string;
+    options?: Options;
+};
