+++ src/compile/tolk/config.ts
@@ -1,0 +1,7 @@
+export type TolkCompilerConfig = {
+    lang: 'tolk';
+    entrypoint: string;
+    optimizationLevel?: number;
+    withStackComments?: boolean;
+    experimentalOptions?: string;
+};
