+++ src/index.ts
@@ -4,7 +4,7 @@ export { NetworkProvider } from './network/NetworkProvider';
 
 export { createNetworkProvider } from './network/createNetworkProvider';
 
-export { compile, CompileOpts, SourceSnapshot, TolkCompileResult, FuncCompileResult, TactCompileResult, CompileResult } from './compile/compile';
+export { compile, CompileOpts, TolkCompileResult, FuncCompileResult, TactCompileResult, CompileResult } from './compile/compile';
 
 export { CompilerConfig, HookParams } from './compile/CompilerConfig';
 
@@ -17,3 +17,4 @@ export { Args, Runner, RunnerContext } from './cli/Runner';
 export { PluginRunner, Plugin } from './config/Plugin';
 export { CustomNetwork } from './config/CustomNetwork';
 export { buildOne, buildAll, buildAllTact } from './build';
+export {SourceSnapshot} from "./compile/SourceSnapshot";
